Fix CRAN pretest failures: clang 21 errors/warnings + Windows byte ambiguity
- krylov.hpp: Remove erroneous .template on topRows(k) runtime calls
- dense_spz_loader.hpp, spz_loader.hpp: Fix member initializer order (-Wreorder)
- speckled_cv.hpp: Explicit cast uint64_t to double (-Wimplicit-const-int-float-conversion)
- spmv.hpp: Fix undefined 'n' variable, guard with #ifdef _OPENMP
- fused_nnls.hpp: Suppress unused 'threads' parameter in non-OpenMP paths
- fit_cv.hpp: Remove unused variables (actual_threads, t_conv_us)
- deflation.hpp: Remove dead k_computed assignment
- sparsepress_v2.hpp: Remove unused private fields, suppress num_threads warning
- fit_cpu.hpp: Fix %ld format with long long args on Windows (use int cast)
- platform.hpp, platform_io.hpp: Fix std::byte vs rpcndr.h byte ambiguity on MinGW
- Makevars.win: Add -Wno-ignored-attributes to suppress Eigen SSE noise

+++ inst/include/FactorNet/nmf/speckled_cv.hpp
@@ -84,7 +84,7 @@ private:
         h = (h ^ (h >> 30)) * 0xbf58476d1ce4e5b9ULL;
         h = (h ^ (h >> 27)) * 0x94d049bb133111ebULL;
         h ^= h >> 31;
-        uint64_t threshold = static_cast<uint64_t>(col_subsample_frac_ * 0xFFFFFFFFFFFFFFFFULL);
+        uint64_t threshold = static_cast<uint64_t>(col_subsample_frac_ * static_cast<double>(0xFFFFFFFFFFFFFFFFULL));
         return h < threshold;
     }
     
@@ -98,7 +98,7 @@ private:
         h = (h ^ (h >> 30)) * 0xbf58476d1ce4e5b9ULL;
         h = (h ^ (h >> 27)) * 0x94d049bb133111ebULL;
         h ^= h >> 31;
-        uint64_t threshold = static_cast<uint64_t>(row_subsample_frac_ * 0xFFFFFFFFFFFFFFFFULL);
+        uint64_t threshold = static_cast<uint64_t>(row_subsample_frac_ * static_cast<double>(0xFFFFFFFFFFFFFFFFULL));
         return h < threshold;
     }
